Add Sass Partials In SMACSS
Add Sass Partials In Scalable and Modular Architecture for CSS (SMACSS)
which divides the styles into base, layout and modules for the time
being.

Base: where the defaults styles will be located such as an element
selectors including its attribute selectors, pseudo-class selectors,
child selectors or sibling selectors. Also this is where the
Sass valuables and mixins are located.

Layout: is where the styling of the global layout of the application
such as the header, footer and any sidebar that goes thought out the
application are located.

Modules: Each individual page has their own Sass partial file in order
to further originated the styles.