Add Clearfix Class To End Each Row
Add Clearfix class to end the row for both the small screens and larger
screens because on small screen there will be two photos per row where
as the lager screens will have three photos per row. The clearfix
function will calculated when to add in the clearfix from Bootstrap.