Add Max Width To Header In Container Row Photos
Add max width to the header, in the container row photos. Because the
width of the photos stop at 500px but the header is still be growing
leaving a whitespace where the photo should be.

This happens when the screen is between the extra small and small
screens. Here the screen is still too small to fit two photos but big
enough to have the photo reach its max width.