Remove Margin From Search Sort Row, Add To All Row
Remove the margin from the top of search & sort row and add to margin
top of each row. Because on smaller screens the whitespace between
search and sorting is too great. By adding to margin top of the row,
this will decrease the need to have a larger margin bottom for each of
the inputs. Thus on smaller screens the whitespace between the inputs
will decrease while keeping the spacing between the photos and sorting
on all screens.